Refactor so that evaluation components are modular and more easily edited.
Evaluation script is also no longer ~500 lines of code lol
